pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-loop"
version = "0.1.0"
description = "Open-source framework for autonomous AI agents powered by Claude Code CLI"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    {name = "Mateusz Rybak"},
]
keywords = ["ai", "agent", "claude", "telegram", "autonomous"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications :: Chat",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "python-telegram-bot>=21.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "apscheduler>=3.10,<4.0",
    "jinja2>=3.1",
]

[project.optional-dependencies]
voice = ["openai-whisper>=20231117"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.4"]

[project.scripts]
agent-loop = "agent_loop.cli:main"

[project.urls]
Homepage = "https://github.com/mateuszrybak/agent-loop"
Documentation = "https://github.com/mateuszrybak/agent-loop/docs"
Repository = "https://github.com/mateuszrybak/agent-loop"
Issues = "https://github.com/mateuszrybak/agent-loop/issues"

[tool.setuptools.packages.find]
include = ["agent_loop*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
